Finalizing JazSharp v1 features.
Implemented Expect(value).ToContain(...)
Implemented Expect(string).ToMatch(regex)
Op: Constrained handled (see below).
base.XXX calls don't fail (spying not supported).
Having native dlls in bin will not cause crash.
Tests for spying on extern methods.

OpCode Constrained is used when calling a struct's method (or a generic parameter which may be a struct) to ensure the this parameter is passed in by reference (so the method can change the value type's properties safely). There was no clean way to handle this so at this time Constrained calls cannot be spied on.